% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_chain.R
\docType{class}
\name{chain}
\alias{chain}
\title{A single langMH chain}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
chain
}
\description{
A single langMH chain
}
\section{Fields}{

\describe{
\item{\code{DL}}{data list}

\item{\code{id}}{identifier of the chain}

\item{\code{seed}}{a starting seed for the chain}

\item{\code{nvars}}{number of variables}

\item{\code{acc_rate}}{cumulative acceptance rate}

\item{\code{x}}{current state of the sampler (most recent draw)}

\item{\code{log.fx}}{log f(x)}

\item{\code{grx}}{gradient at x}

\item{\code{draws}}{all collected draws}

\item{\code{acc}}{vector of whether an iteration was accepted or not}

\item{\code{iter}}{iteration ids for each draw}

\item{\code{logpost}}{log posterior for each draw}

\item{\code{last_iter}}{iteration id of most recent draw}

\item{\code{last_draw}}{draw id of most recent draw}

\item{\code{varnames}}{vector of variable names for draws}

\item{\code{e1,e2,A1,delta}}{adaptation parameters}

\item{\code{mu,G,sig,norm.grx}}{internal adaptation parameters}

\item{\code{opt_acc}}{optimal acceptance rate}
}}
\keyword{datasets}

