% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dMVN}
\alias{dMVN}
\title{dMVN}
\usage{
dMVN(X_, mu_, L_, isPrec)
}
\arguments{
  \item{X_}{matrix} \item{mu_}{vector} \item{L_}{lower chol
  of cov or prec matrix} \item{isPrec}{covariance or precision
  matrix?}
}
\value{
Numeric vector
}

