% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads-utilities.R
\name{getcreatives}
\alias{getcreatives}
\title{Generate creatives renewal/replacement/addition. Currently only identifies (0,1) whether}
\usage{
getcreatives(category, brands.adv, fweek, T)
}
\arguments{
  \item{category}{Three letter category acronym (being one
  of dpp, fti, lld, ptw, tti)} \item{brands.adv}{Which brands
  advertised} \item{fweek}{Integer value for first week to
  start analysis} \item{T}{Integer value for number of weeks
  to include from and inclusive of fweek}
}
\value{
T x Jb matrix with integer corresponding to number of new
creatives added that week
}
\description{
Generate creatives renewal/replacement/addition. Currently
only identifies (0,1) whether
}
\examples{
getcreatives(category, brands.adv, Jb, fweek, T) - called from within function mcmod()
}

