% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads-utilities.R
\name{mcmodf}
\alias{mcmodf}
\title{mcmodf
Create data files for analysis from raw IRI and TNS data files.}
\usage{
mcmodf(data.name = "dpp", brands.to_keep = c("HUGGIES", "PAMPERS", "LUVS",
  "PRIVATE LABEL"), covv = c("avprc"), covnv = c("fracfnp", "fracdnp",
  "fracdist", "numproducts"), T = 226, N = 42, fweek = 1200,
  aggregated = FALSE, max.distance = 0.2, summary.only = FALSE)
}
\arguments{
  \item{data.name}{Three letter category acronym (being one
  of dpp, fti, lld, ptw, tti)} \item{brands.to_keep}{List of
  brands corresponding to the Y matrix (being volume sales)}
  \item{covv}{Character vector for time varying covariates,
  being one of 'avprc', 'lavprc', and others} \item{T}{Integer
  value for number of weeks to include from and inclusive
  of fweek} \item{N}{Integer value for number of markets to
  include (constructs from market_list being an alphabetic
  list)} \item{fweek}{Integer value for first week to start
  analysis} \item{aggregated}{(Defaults to FALSE) Boolean flag
  to indicate generation of aggregated (all US) versus city
  level data}
}
\value{
mcmod object being a list with the following elements:
list(dimensions=dimensions,Y = Y,CM = CMl, E = El, A = Al, X
= Xl, F1 = F1l, F2 = F2)
}
\description{
mcmodf Create data files for analysis from raw IRI and TNS
data files.
}
\examples{
mcmod() - Runs default with "dpp"
mcmod(data.name = "dpp", brands.to_keep = c('HUGGIES','PAMPERS','LUVS','PL'), covv = c("avprc"), covnv = c("fracfnp","fracdnp","fracdist","numproducts"), T = 226,  N = 42, fweek = 1200, aggregated = FALSE)
}

