% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads-utilities.R
\name{rmvMN}
\alias{rmvMN}
\title{rmvMN
Draw samples from a matrix normal using Cholesky decomposition.}
\usage{
rmvMN(ndraws, M = rep(0, nrow(S) * ncol(C)), C, S)
}
\arguments{
  \item{ndraws}{Number of draws} \item{M}{matrix mean (must
  by dimension nrow(S) x nrow(C))} \item{C}{covariance
  matrix representing columns} \item{S}{covariance matrix
  representing rows}
}
\value{
ndraws matrices each of dimension S x C of random normal
distribution
}
\description{
rmvMN Draw samples from a matrix normal using Cholesky
decomposition.
}
\examples{
rmvMN(1, C, S)
}

