% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads-utilities.R
\name{trim_characters}
\alias{trim_characters}
\title{Trim white space.}
\usage{
trim_characters(DT)
}
\arguments{
  \item{DT}{Data table name}
}
\value{
A new data table with trimmed characters
}
\description{
Trim characters recursively in a data table.
}
\examples{
b <- trim_characters(a)
}

