% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_chain.R
\docType{class}
\name{chain}
\alias{chain}
\title{A single langMH chain}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
chain
}
\description{
A single langMH chain
}
\section{Fields}{

\describe{
\item{\code{DL}}{data list}

\item{\code{id}}{identifier of the chain}

\item{\code{seed}}{a starting seed for the chain}

\item{\code{nvars}}{number of variables}

\item{\code{acc_rate}}{cumulative acceptance rate}

\item{\code{x}}{current state of the sampler}

\item{\code{log.fx}}{log f(x)}

\item{\code{grx}}{gradient at x}
}}
\keyword{datasets}

